Create LICENSE
This should be compatible with incorporating any of the contents into another project at a later time, as long as our copyright is noted.

I have not yet committed any files with other pre-existing copyrights. For simplicity, we may choose to keep those in separate Zenodo archives or other repositories.